requirements.txt
arrow==0.10.0
boto==2.48.0
certifi==2017.7.27.1
chardet==3.0.4
colorlog==3.1.0
contextdecorator==0.10.0
cycler==0.10.0
future==0.16.0
idna==2.5
iso8601==0.1.11
jsonschema==2.6.0
matplotlib==1.5.1
numpy==1.13.1
parsedatetime==2.4
ply==3.10
protobuf==3.2.0
pyformance==0.3.2
pyparsing==2.1.10
PyStaticConfiguration==0.10.3
pysubnettree==0.26
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.18.2
setuptools==36.2.7
signalfx==1.0.16
simplejson==3.11.1
six==1.10.0
sortedcontainers==1.5.7
sseclient-py==1.4
thriftpy==0.3.9
urllib3==1.22
ws4py==0.3.4
yelp-clog==2.10.0
yelp-lib==13.1.2
yelp-servlib==5.6.0
